move ActionResult to models file
